pyproject.toml
[project]
name = "ptl-buddy-server"
version = "0.0.0"
requires-python = ">=3.14"
dependencies = [
	"Flask<4",
	"Flask-SQLAlchemy-Lite<0.3",
	"gunicorn<26",
	"ldap3<3",
	"siphash<0.1",
]

[dependency-groups]
dev = [
	"python-dotenv<2",
	"ruff<0.16",
]

[project.scripts]
ptl-buddy-server = "ptl_buddy_server.prod:application.run"

[build-system]
requires = ["uv_build<0.12"]
build-backend = "uv_build"

[tool.ruff]
lint.select = ["ALL"]
lint.ignore = [
	"D1",  # undocumented code
	"ANN401",  # typing.Any
	"D203", "D213",  # conflict with other checks
]
line-length = 120
format.quote-style = "single"
lint.flake8-comprehensions.allow-dict-calls-with-keyword-arguments = true
lint.flake8-quotes.inline-quotes = "single"
